requirements.txt
#redis_cache
Django==3.2
django-import-export
